check arcs

+++ automato_periodo.py
@@ -7,6 +7,7 @@ import zipfile
 import random
 from ftplib import FTP, error_perm
 import socket
+import time
 
 # Obter o diretório "Documentos" do usuário atual
 pasta_documentos = os.path.expanduser("~" + os.sep + "Documents")
@@ -44,7 +45,13 @@ def log_message(message):
     with open(nome_arquivo, 'a') as log_file:
         log_file.write(f"[{timestamp}] {message}\n")
     print(message)
-    
+
+def _existe(file_name, folder_path):
+    # Cria o caminho completo do arquivo
+    file_path = os.path.join(folder_path, file_name)
+    # Verifica se o caminho é um arquivo existente
+    return os.path.isfile(file_path)
+
 data_log = datetime.now()
 
 log_message('----Iniciando login na BISP----')
@@ -105,7 +112,7 @@ try:
         
         query_1 = """
             -- 1 - REDS_RAT - OK
-            SELECT	
+            SELECT  
                 OCO.numero_ocorrencia AS "RAT.NUM_ATIVIDADE",
                 OCO.natureza_codigo AS "NAT.CODIGO",
                 OCO.natureza_descricao AS "NAT.DESCRICAO",
@@ -150,25 +157,37 @@ try:
             ORDER BY OCO.data_hora_fato;
         """.format(data_inicial, data_final)
 
-        log_message('----Iniciando processamento da query.----')
-        cursor.execute(query_1)
-        log_message('----Query 1 processada com sucesso!----')
+        log_message('----Iniciando processamento da query 1----')
 
-        df = pd.DataFrame(cursor.fetchall(), columns=[col[0] for col in cursor.description])
-
-        # Converte as colunas em uppercase
-        df.columns = [col.upper() for col in df.columns]
-        
         arquivo_csv = f"CARNAVAL_REDS_RAT_{data_consulta}_{data_consulta}.csv"
         arquivo_zip = f"CARNAVAL_REDS_RAT_{data_consulta}_{data_consulta}.zip"
-        df.to_csv(arquivo_csv, index=False, sep='|', encoding='utf-8')
-        with zipfile.ZipFile(arquivo_zip, 'w') as zipf:
-            zipf.write(arquivo_csv)
+
+        # Confere se o arquivo já existe
+        file_name = arquivo_zip
+        folder_path = rat_dia
+
+        if _existe(file_name, folder_path):
+            print(f'O arquivo {arquivo_zip} já existe.')
+            time.sleep(1)
+        else:
+            cursor.execute(query_1)
+
+            df = pd.DataFrame(cursor.fetchall(), columns=[col[0] for col in cursor.description])
+
+            # Converte as colunas em uppercase
+            df.columns = [col.upper() for col in df.columns]
+            
+            df.to_csv(arquivo_csv, index=False, sep='|', encoding='utf-8')
+            with zipfile.ZipFile(arquivo_zip, 'w') as zipf:
+                zipf.write(arquivo_csv)
+            
+            log_message('----Query 1 processada com sucesso!----')
+            os.remove(arquivo_csv) 
         
-        os.remove(arquivo_csv)
+       
         query_2 = """
             -- 2 -REDS_RAT_EFETIVOS - OK
-            SELECT	
+            SELECT  
                 OCO.numero_ocorrencia AS 'NUM_ATIVIDADE',
                 OCO.digitador_matricula AS 'NUM_MATRICULA',
                 OCO.digitador_nome AS 'NOME',
@@ -177,28 +196,41 @@ try:
                 OCO.unidade_responsavel_registro_nome AS 'NOM_UNIDADE'
             FROM
                 db_bisp_reds_reporting.tb_ocorrencia OCO
-            WHERE 1=1
+            WHERE 1=16
             AND OCO.nome_tipo_relatorio = 'RAT'
             AND OCO.ind_estado IN ('R', 'F')
             AND OCO.data_hora_fato IS NOT NULL
             AND OCO.data_hora_alteracao BETWEEN '{}' AND '{}'
             ORDER BY OCO.data_hora_fato;
         """.format(data_inicial, data_final)
-        cursor.execute(query_2)
-        log_message('----Query 2 processada com sucesso!----')
 
-        df = pd.DataFrame(cursor.fetchall(), columns=[col[0] for col in cursor.description])
-
-        # Converte as colunas em uppercase
-        df.columns = [col.upper() for col in df.columns]
+        log_message('----Iniciando processamento da query 2----')
 
         arquivo_csv = f"CARNAVAL_REDS_RAT_EFETIVOS_{data_consulta}_{data_consulta}.csv"
         arquivo_zip = f"CARNAVAL_REDS_RAT_EFETIVOS_{data_consulta}_{data_consulta}.zip"
-        df.to_csv(arquivo_csv, index=False, sep='|', encoding='utf-8')
-        with zipfile.ZipFile(arquivo_zip, 'w') as zipf:
-            zipf.write(arquivo_csv)
         
-        os.remove(arquivo_csv)
+        # Confere se o arquivo já existe
+        file_name = arquivo_zip
+        folder_path = rat_dia
+
+        if _existe(file_name, folder_path):
+            print(f'O arquivo {arquivo_zip} já existe.')
+            time.sleep(1)
+        else:
+            cursor.execute(query_2)
+
+            df = pd.DataFrame(cursor.fetchall(), columns=[col[0] for col in cursor.description])
+
+            # Converte as colunas em uppercase
+            df.columns = [col.upper() for col in df.columns]
+            
+            df.to_csv(arquivo_csv, index=False, sep='|', encoding='utf-8')
+            with zipfile.ZipFile(arquivo_zip, 'w') as zipf:
+                zipf.write(arquivo_csv)
+            
+            log_message('----Query 2 processada com sucesso!----')
+            os.remove(arquivo_csv) 
+                
         
         query_3 = """
         -- 3 - REDS_RAT_produtividade - OK
@@ -217,22 +249,33 @@ try:
             AND PROD.quantidade <> 0
             ORDER BY PROD.data_hora_fato;
         """.format(data_inicial, data_final)
-        cursor.execute(query_3)
-        log_message('----Query 3 processada com sucesso!----')
-
-        df = pd.DataFrame(cursor.fetchall(), columns=[col[0] for col in cursor.description])
-
-        # Converte as colunas em uppercase
-        df.columns = [col.upper() for col in df.columns]
+        log_message('----Iniciando processamento da query 3----')
 
         arquivo_csv = f"CARNAVAL_REDS_RAT_Produtividade_{data_consulta}_{data_consulta}.csv"
         arquivo_zip = f"CARNAVAL_REDS_RAT_Produtividade_{data_consulta}_{data_consulta}.zip"
-        df.to_csv(arquivo_csv, index=False, sep='|', encoding='utf-8')
-        with zipfile.ZipFile(arquivo_zip, 'w') as zipf:
-            zipf.write(arquivo_csv)
         
-        os.remove(arquivo_csv)
-        
+        # Confere se o arquivo já existe
+        file_name = arquivo_zip
+        folder_path = rat_dia
+
+        if _existe(file_name, folder_path):
+            print(f'O arquivo {arquivo_zip} já existe.')
+            time.sleep(1)
+        else:
+            cursor.execute(query_3)
+
+            df = pd.DataFrame(cursor.fetchall(), columns=[col[0] for col in cursor.description])
+
+            # Converte as colunas em uppercase
+            df.columns = [col.upper() for col in df.columns]
+            
+            df.to_csv(arquivo_csv, index=False, sep='|', encoding='utf-8')
+            with zipfile.ZipFile(arquivo_zip, 'w') as zipf:
+                zipf.write(arquivo_csv)
+            
+            log_message('----Query 3 processada com sucesso!----')
+            os.remove(arquivo_csv) 
+
         query_4 = """
             -- 4 - REDS_RAT_VIATURAS - OK
             SELECT	
@@ -252,22 +295,34 @@ try:
             AND VTR.data_hora_fato BETWEEN '{}' AND '{}'
             ORDER BY VTR.data_hora_fato;
         """.format(data_inicial, data_final)
-        cursor.execute(query_4)
-        df = pd.DataFrame(cursor.fetchall(), columns=[col[0] for col in cursor.description])
-
-        # Converte as colunas em uppercase
-        df.columns = [col.upper() for col in df.columns]
+        log_message('----Iniciando processamento da query 4----')
 
         arquivo_csv = f"CARNAVAL_REDS_RAT_VIATURAS_{data_consulta}_{data_consulta}.csv"
         arquivo_zip = f"CARNAVAL_REDS_RAT_VIATURAS_{data_consulta}_{data_consulta}.zip"
-        df.to_csv(arquivo_csv, index=False, sep='|', encoding='utf-8')
-        with zipfile.ZipFile(arquivo_zip, 'w') as zipf:
-            zipf.write(arquivo_csv)
-            
-        log_message('----Query 4 processada com sucesso!----')
-        
-        os.remove(arquivo_csv)
 
+        # Confere se o arquivo já existe
+        file_name = arquivo_zip
+        folder_path = rat_dia
+
+        if _existe(file_name, folder_path):
+            print(f'O arquivo {arquivo_zip} já existe.')
+            time.sleep(1)
+        else:
+            cursor.execute(query_4)
+
+            df = pd.DataFrame(cursor.fetchall(), columns=[col[0] for col in cursor.description])
+
+            # Converte as colunas em uppercase
+            df.columns = [col.upper() for col in df.columns]
+            
+            df.to_csv(arquivo_csv, index=False, sep='|', encoding='utf-8')
+            with zipfile.ZipFile(arquivo_zip, 'w') as zipf:
+                zipf.write(arquivo_csv)
+            
+            log_message('----Query 4 processada com sucesso!----')
+            os.remove(arquivo_csv) 
+
+        # Muda diretório de gravação para BOS
         os.chdir(bos_dia)
 
         query_5 = """
@@ -315,21 +370,33 @@ try:
             AND OCO.data_hora_alteracao BETWEEN '{}' AND '{}'
             ORDER BY OCO.data_hora_fato;
         """.format(data_inicial, data_final)
-        cursor.execute(query_5)
-        log_message('----Query 5 processada com sucesso!----')
-
-        df = pd.DataFrame(cursor.fetchall(), columns=[col[0] for col in cursor.description])
-
-        # Converte as colunas em uppercase
-        df.columns = [col.upper() for col in df.columns]
+        log_message('----Iniciando processamento da query 5----')
 
         arquivo_csv = f"CARNAVAL_REDS_BOS_{data_consulta}_{data_consulta}.csv"
         arquivo_zip = f"CARNAVAL_REDS_BOS_{data_consulta}_{data_consulta}.zip"
-        df.to_csv(arquivo_csv, index=False, sep='|', encoding='utf-8')
-        with zipfile.ZipFile(arquivo_zip, 'w') as zipf:
-            zipf.write(arquivo_csv)
         
-        os.remove(arquivo_csv)
+        # Confere se o arquivo já existe
+        file_name = arquivo_zip
+        folder_path = rat_dia
+
+        if _existe(file_name, folder_path):
+            print(f'O arquivo {arquivo_zip} já existe.')
+            time.sleep(1)
+        else:
+            cursor.execute(query_5)
+
+            df = pd.DataFrame(cursor.fetchall(), columns=[col[0] for col in cursor.description])
+
+            # Converte as colunas em uppercase
+            df.columns = [col.upper() for col in df.columns]
+            
+            df.to_csv(arquivo_csv, index=False, sep='|', encoding='utf-8')
+            with zipfile.ZipFile(arquivo_zip, 'w') as zipf:
+                zipf.write(arquivo_csv)
+            
+            log_message('----Query 5 processada com sucesso!----')
+            os.remove(arquivo_csv) 
+
         query_6 = """
             -- 6 - REDS_BOS_EFETIVOS - OK
             SELECT	
@@ -348,21 +415,33 @@ try:
             AND OCO.data_hora_alteracao BETWEEN '{}' AND '{}'
             ORDER BY OCO.data_hora_fato;
         """.format(data_inicial, data_final)
-        cursor.execute(query_6)
-        log_message('----Query 6 processada com sucesso!----')
-
-        df = pd.DataFrame(cursor.fetchall(), columns=[col[0] for col in cursor.description])
-
-        # Converte as colunas em uppercase
-        df.columns = [col.upper() for col in df.columns]
+        
+        log_message('----Iniciando processamento da query 6----')
 
         arquivo_csv = f"CARNAVAL_REDS_BOS_EFETIVOS_{data_consulta}_{data_consulta}.csv"
         arquivo_zip = f"CARNAVAL_REDS_BOS_EFETIVOS_{data_consulta}_{data_consulta}.zip"
-        df.to_csv(arquivo_csv, index=False, sep='|', encoding='utf-8')
-        with zipfile.ZipFile(arquivo_zip, 'w') as zipf:
-            zipf.write(arquivo_csv)
-        
-        os.remove(arquivo_csv)
+              
+        # Confere se o arquivo já existe
+        file_name = arquivo_zip
+        folder_path = rat_dia
+
+        if _existe(file_name, folder_path):
+            print(f'O arquivo {arquivo_zip} já existe.')
+            time.sleep(1)
+        else:
+            cursor.execute(query_6)
+
+            df = pd.DataFrame(cursor.fetchall(), columns=[col[0] for col in cursor.description])
+
+            # Converte as colunas em uppercase
+            df.columns = [col.upper() for col in df.columns]
+            
+            df.to_csv(arquivo_csv, index=False, sep='|', encoding='utf-8')
+            with zipfile.ZipFile(arquivo_zip, 'w') as zipf:
+                zipf.write(arquivo_csv)
+            
+            log_message('----Query 6 processada com sucesso!----')
+            os.remove(arquivo_csv) 
         
         query_7 = """
             -- 7 - REDS_BOS_ENVOLVIDO - OK
@@ -395,21 +474,33 @@ try:
             AND ENV.nome_completo_envolvido <> ''
             ORDER BY OCO.data_hora_fato;
         """.format(data_inicial, data_final)
-        cursor.execute(query_7)
-        log_message('----Query 7 processada com sucesso!----')
 
-        df = pd.DataFrame(cursor.fetchall(), columns=[col[0] for col in cursor.description])
-
-        # Converte as colunas em uppercase
-        df.columns = [col.upper() for col in df.columns]
+        log_message('----Iniciando processamento da query 7----')
 
         arquivo_csv = f"CARNAVAL_REDS_BOS_ENVOLVIDO_{data_consulta}_{data_consulta}.csv"
         arquivo_zip = f"CARNAVAL_REDS_BOS_ENVOLVIDO_{data_consulta}_{data_consulta}.zip"
-        df.to_csv(arquivo_csv, index=False, sep='|', encoding='utf-8')
-        with zipfile.ZipFile(arquivo_zip, 'w') as zipf:
-            zipf.write(arquivo_csv)
-        
-        os.remove(arquivo_csv)
+
+        # Confere se o arquivo já existe
+        file_name = arquivo_zip
+        folder_path = rat_dia
+
+        if _existe(file_name, folder_path):
+            print(f'O arquivo {arquivo_zip} já existe.')
+            time.sleep(1)
+        else:
+            cursor.execute(query_7)
+
+            df = pd.DataFrame(cursor.fetchall(), columns=[col[0] for col in cursor.description])
+
+            # Converte as colunas em uppercase
+            df.columns = [col.upper() for col in df.columns]
+            
+            df.to_csv(arquivo_csv, index=False, sep='|', encoding='utf-8')
+            with zipfile.ZipFile(arquivo_zip, 'w') as zipf:
+                zipf.write(arquivo_csv)
+            
+            log_message('----Query 7 processada com sucesso!----')
+            os.remove(arquivo_csv) 
         
         query_8 = """
             -- 8 - REDS_BOS_VIATURAS - OK
@@ -430,23 +521,35 @@ try:
             AND VTR.data_hora_fato BETWEEN '{}' AND '{}'
             ORDER BY VTR.data_hora_fato;
         """.format(data_inicial, data_final)
-        cursor.execute(query_8)
-        log_message('----Query 8 processada com sucesso!----')
 
-        df = pd.DataFrame(cursor.fetchall(), columns=[col[0] for col in cursor.description])
-
-        # Converte as colunas em uppercase
-        df.columns = [col.upper() for col in df.columns]
+        log_message('----Iniciando processamento da query 8----')
 
         arquivo_csv = f"CARNAVAL_REDS_BOS_VIATURAS_{data_consulta}_{data_consulta}.csv"
         arquivo_zip = f"CARNAVAL_REDS_BOS_VIATURAS_{data_consulta}_{data_consulta}.zip"
-        df.to_csv(arquivo_csv, index=False, sep='|', encoding='utf-8')
-        with zipfile.ZipFile(arquivo_zip, 'w') as zipf:
-            zipf.write(arquivo_csv)
-        data_atual = data_atual + timedelta(days=1)
         
+         # Confere se o arquivo já existe
+        file_name = arquivo_zip
+        folder_path = rat_dia
+
+        if _existe(file_name, folder_path):
+            print(f'O arquivo {arquivo_zip} já existe.')
+            time.sleep(1)
+        else:
+            cursor.execute(query_8)
+
+            df = pd.DataFrame(cursor.fetchall(), columns=[col[0] for col in cursor.description])
+
+            # Converte as colunas em uppercase
+            df.columns = [col.upper() for col in df.columns]
+            
+            df.to_csv(arquivo_csv, index=False, sep='|', encoding='utf-8')
+            with zipfile.ZipFile(arquivo_zip, 'w') as zipf:
+                zipf.write(arquivo_csv)
+            os.remove(arquivo_csv) 
+            log_message('----Query 8 processada com sucesso!----')
+
+        data_atual = data_atual + timedelta(days=1)
         log_message(f'Reiniciando protocolo para o dia {data_atual}')
-        os.remove(arquivo_csv)
             
 except ImpalaError as e:
     log_message('***** Erro de autenticação: {e} *****')
